requirements.txt
Django<2.2
line-bot-sdk
gunicorn
psycopg2
social-auth-app-django
mysqlclient
whitenoise